--- targets/ChibiOS/RP_PICO_W_RP2040/sys_dev_adc_native_target.h
@@ -1,7 +1,0 @@
-//
-// Copyright (c) .NET Foundation and Contributors
-// See LICENSE file in the project root for full license information.
-//
-
-// Shared RP2040 implementation
-#include "../_RP2040/sys_dev_adc_native_target.h"
