+++ targets/ChibiOS/_nanoCLR/System.Device.Adc/sys_dev_adc_native_target.h
@@ -9,6 +9,18 @@
 #include <hal.h>
 #include <sys_dev_adc_native.h>
 
+#if defined(RP_ADC_USE_ADC1)
+
+// RP2040 ADC: single ADC with static channel map.
+typedef struct
+{
+    uint8_t adcIndex;
+    uint8_t gpio; // GPIO number (26-29) or 0xFF for internal channels.
+    uint32_t adcChannel;
+} NF_PAL_ADC_PORT_PIN_CHANNEL;
+
+#else
+
 typedef struct
 {
     uint8_t adcIndex;
@@ -17,10 +29,14 @@ typedef struct
     uint32_t adcChannel;
 } NF_PAL_ADC_PORT_PIN_CHANNEL;
 
+#endif
+
 extern const NF_PAL_ADC_PORT_PIN_CHANNEL AdcPortPinConfig[];
 extern const int AdcChannelCount;
 
+#if !defined(RP_ADC_USE_ADC1)
 extern NF_PAL_ADC_PORT_PIN_CHANNEL *RuntimeAdcPortPinConfig;
 extern uint8_t RuntimeAdcChannelCount;
+#endif
 
 #endif // SYS_DEV_ADC_NATIVE_TARGET_H
